harpoon-term: add support for floating windows
Terminals can now be opened in floating windows.

toggle_floating_window(idx) can be used to toggle a floating window
containing a terminal.

The goto_terminal(idx) and send_command(idx, cmd) functions will operate
on floating windows if a terminal is open using a floating window.

The dedicated send_command_to_floating_window(idx, cmd) function can be
used to ensure that the command is sent to a floating window that is
toggled open when the command is sent.